Add face upload/training
Also, FaceBox and FaceBoxCollection now is refactored into its own file
